+++ plugins/fabrik_element/emundus_fileupload/emundus_fileupload.php
@@ -119,7 +119,7 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
                     $sizeMax = ($postSize >= $iniSize) ? $iniSize : $postSize;
 
                     if (!empty($fileName)) {
-                        $insert[] = $user . ' , ' . $db->quote($fnum) . ' , ' . $cid . ' , ' . $attachId . ' , ' . $db->quote($fileName) . ' , ' . '0' . ' , ' . '1';
+                        $insert[] = $user . ' , ' . $db->quote($fnum) . ' , ' . $cid . ' , ' . $attachId . ' , ' . $db->quote($fileName) . ' , ' . '1' . ' , ' . '1' . ' , ' . $db->quote($file['name']);
                     }
 
 
@@ -138,7 +138,7 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
 
                     $sizeMax = $this->formatBytes($sizeMax);
 
-                    $result[] = array('size' => $size, 'ext' => $ext, 'nbMax' => $fileLimitObtained, 'filename' => $fileName, 'target' => $target, 'nbAttachment' => $nbAttachment, 'encrypt' => $encrypt, 'maxSize' => $sizeMax);
+                    $result[] = array('size' => $size, 'ext' => $ext, 'nbMax' => $fileLimitObtained, 'filename' => $fileName,'local_filename' => $file['name'], 'target' => $target, 'nbAttachment' => $nbAttachment, 'encrypt' => $encrypt, 'maxSize' => $sizeMax);
                     if ($size === false || $fileLimitObtained === true) {
                         echo json_encode($result);
                         return true;
@@ -146,7 +146,7 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
                 } else {
                     $size = true;
                     $ext = false;
-                    $result[] = array('size' => $size, 'ext' => $ext, 'filename' => $fileName, 'target' => $target, 'nbAttachment' => $nbAttachment);
+                    $result[] = array('size' => $size, 'ext' => $ext, 'filename' => $fileName,'local_filename' => $file['name'], 'target' => $target, 'nbAttachment' => $nbAttachment);
                     echo json_encode($result);
                     return true;
                 }
@@ -256,13 +256,18 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
         $result = array('limitObtained' => $nbMaxFile<=sizeof($uploadResult));
 
         foreach ($uploadResult as $upload) {
+	        $fileName = '';
+	        $local_fileName = '';
             if (!empty($upload->filename)) {
-                $fileName = $upload->filename;
+	            $fileName = $upload->filename;
+	            $local_fileName = $upload->filename;
+				if(!empty($upload->local_filename)){
+					$local_fileName = $upload->local_filename;
+				}
             }
 
             $target = '/images'.DS.'emundus'.DS.'files'.DS.$user.DS.$fileName;
-            $result['files'][] = array('filename' => $fileName, 'target' => $target);
-
+            $result['files'][] = array('filename' => $fileName,'local_filename' => $local_fileName, 'target' => $target, 'can_be_deleted' => $upload->can_be_deleted, 'can_be_viewed' => $upload->can_be_viewed);
         }
 
         echo json_encode($result);
@@ -276,48 +281,45 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
      * @throws Exception
      */
     public function onAjax_delete() {
+		$current_user = JFactory::getSession()->get('emundusUser');
 
-        $jinput = $this->app->input;
-        $current_user = JFactory::getSession()->get('emundusUser');
+	    $result = array('status' => false);
 
-        $fileName = $jinput->post->get('filename');
-        $attachId = $jinput->post->get('attachId');
+        if (EmundusHelperAccess::isApplicant($current_user->id)) {
+            $jinput = $this->app->input;
+            $fileName = $jinput->post->get('filename');
+            $attachId = $jinput->post->get('attachId');
+            $fnum = $jinput->post->get('fnum');
+            $user = (int)substr($fnum, -7);
+            $cid = $this->getCampaignId($fnum);
+            $uploadResult = $this->getUploads($attachId, $user, $cid, $fnum);
 
-        if (!EmundusHelperAccess::isApplicant($current_user->id)) {
-            return false;
+            if (!empty($uploadResult)) {
+                $target = $this->getPath($user, $fileName);
+
+                if (file_exists($target)) {
+                    unlink($target);
+                }
+
+	            $result['status'] = $this->deleteFile($fileName, $fnum, $cid, $attachId);
+
+                if ($result['status']) {
+                    // track the LOGS (ATTACHMENT_DELETE)
+                    $user = JFactory::getSession()->get('emundusUser'); # logged user #
+
+                    require_once(JPATH_SITE . '/components/com_emundus/models/files.php');
+                    $mFile = new EmundusModelFiles();
+                    $applicant_id = ($mFile->getFnumInfos($fnum))['applicant_id'];
+
+                    if (!empty($applicant_id)) {
+                        require_once(JPATH_SITE . '/components/com_emundus/models/logs.php');
+                        EmundusModelLogs::log($user->id, $applicant_id, $fnum, 4, 'd', 'COM_EMUNDUS_ACCESS_ATTACHMENT_DELETE');
+                    }
+                }
+            }
         }
 
-        $fnum = $jinput->post->get('fnum');
-        $user = (int)substr($fnum, -7);
-
-        $cid = $this->getCampaignId($fnum);
-        $uploadResult = $this->getUploads($attachId, $user, $cid, $fnum);
-        $target = $this->getPath($user, $fileName);
-
-        if (file_exists($target) && !empty($uploadResult)) {
-            unlink($target);
-            $this->deleteFile($fileName, $fnum, $cid, $attachId);
-            $status = true;
-        }
-        if (!file_exists($target) && !empty($uploadResult)) {
-            $this->deleteFile($fileName, $fnum, $cid, $attachId);
-            $status = true;
-        }
-        if (!file_exists($target) && empty($uploadResult)) {
-            $status = false;
-        }
-
-        // track the LOGS (ATTACHMENT_DELETE)
-        require_once(JPATH_SITE.DS.'components'.DS.'com_emundus'.DS.'models'.DS.'logs.php');
-        $user = JFactory::getSession()->get('emundusUser'); # logged user #
-
-        require_once(JPATH_SITE.DS.'components'.DS.'com_emundus'.DS.'models'.DS.'files.php');
-        $mFile = new EmundusModelFiles();
-        $applicant_id = ($mFile->getFnumInfos($fnum))['applicant_id'];
-
-        EmundusModelLogs::log($user->id, $applicant_id, $fnum, 4, 'd', 'COM_EMUNDUS_ACCESS_ATTACHMENT_DELETE');
-
-        echo json_encode(['status' => $status]);
+        echo json_encode($result);
         return true;
     }
 
@@ -412,7 +414,7 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
         $db = JFactory::getDBO();
 
         $query = $db->getQuery(true);
-        $query->select(array($db->quoteName('id'),$db->quoteName('filename')))
+        $query->select(array($db->quoteName('id'),$db->quoteName('filename'),$db->quoteName('local_filename'),$db->quoteName('can_be_deleted'),$db->quoteName('can_be_viewed')))
             ->from($db->quoteName('#__emundus_uploads'))
             ->where($db->quoteName('attachment_id') . ' = ' . $attachId . ' AND ' . $db->quoteName('fnum') . ' LIKE ' . $db->quote($fnum));
         $db->setQuery($query);
@@ -644,6 +646,8 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
         JText::script('PLG_ELEMENT_FIELD_CANCEL');
         JText::script('PLG_ELEMENT_FIELD_DELETE');
         JText::script('PLG_ELEMENT_FIELD_DELETE_TEXT');
+        JText::script('PLG_ELEMENT_FIELD_DELETE_FAILED');
+        JText::script('PLG_ELEMENT_FIELD_DELETE_TEXT_FAILED');
         JText::script('PLG_ELEMENT_FIELD_ACCESS');
         JText::script('PLG_ELEMENT_FIELD_UPLOAD');
         JText::script('PLG_ELEMENT_FILEUPLOAD_UPLOADED_FILES');
@@ -753,7 +757,7 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
 
             $db = JFactory::getDBO();
             $query = $db->getQuery(true);
-            $columns = array('user_id', 'fnum', 'campaign_id', 'attachment_id', 'filename', 'can_be_deleted', 'can_be_viewed');
+            $columns = array('user_id', 'fnum', 'campaign_id', 'attachment_id', 'filename', 'can_be_deleted', 'can_be_viewed','local_filename');
 
             $query->insert($db->quoteName('#__emundus_uploads'))
                 ->columns($db->quoteName($columns))
@@ -803,20 +807,29 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
      *
      * @throws Exception
      */
-    public function deleteFile($fileName,$fnum, $cid, $attachId) {
-        $db = JFactory::getDBO();
-        $query = $db->getQuery(true);
+    public function deleteFile($fileName, $fnum, $cid, $attachId) {
+        $deleted = false;
 
-        $query->delete($db->quoteName('#__emundus_uploads'))
-            ->where($db->quoteName('filename'). " LIKE " . $db->quote($fileName) . ' AND ' .$db->quoteName('campaign_id') . ' = ' . $cid . " AND " . $db->quoteName('attachment_id') . " = " . $attachId . " AND " . $db->quoteName('fnum') . " LIKE " . $db->quote($fnum));
-        $db->setQuery($query);
+        if (!empty($fnum) && !empty($fileName) && !empty($attachId) && !empty($cid)) {
+            $db = JFactory::getDBO();
+            $query = $db->getQuery(true);
 
-        try {
-            $db->execute();
-        } catch (Exception $e) {
-            JFactory::getApplication()->enqueueMessage('Problème survenu à la suppression des fichiers', 'message');
+            $query->delete($db->quoteName('#__emundus_uploads'))
+                ->where($db->quoteName('filename'). ' LIKE ' . $db->quote($fileName))
+                ->andWhere( $db->quoteName('campaign_id') . ' = ' . $cid)
+                ->andWhere( $db->quoteName('attachment_id') . ' = ' . $attachId)
+                ->andWhere( $db->quoteName('fnum') . ' LIKE ' . $db->quote($fnum));
+
+            try {
+                $db->setQuery($query);
+                $deleted = $db->execute();
+            } catch (Exception $e) {
+                JLog::add("Failed to delete file for fnum $fnum, filename $fileName, campaign $cid, attachment $attachId : " . $e->getMessage(), JLog::ERROR, 'com_emundus.error');
+                JFactory::getApplication()->enqueueMessage('Problème survenu à la suppression des fichiers', 'message');
+            }
         }
 
+        return $deleted;
     }
